IDE-686 Tab settings issue
Fixed IDE-686

+++ wlib/SourceView.cpp
@@ -402,7 +402,7 @@ void CSourceCtrl::DoInit()
 
     InitColors(m_langRef);
 
-    SetTabWidth(GetIConfig(QUERYBUILDER_CFG)->Get(GLOBAL_TAB_WIDTH));
+    SetTabWidth((int)GetIConfig(QUERYBUILDER_CFG)->Get(GLOBAL_TAB_WIDTH));
     SetUseTabs(!(bool)GetIConfig(QUERYBUILDER_CFG)->Get(GLOBAL_TAB_USESPACES));
 
     if (GetIConfig(QUERYBUILDER_CFG)->Get(GLOBAL_LINENO))
